navbar desktop and mobile split into 2 components, creation mediaquery hook to handle which one to display, code needs cleaning
