+++ .gitignore
@@ -24,4 +24,4 @@ yarn-error.log*
 
 starter-code/
 instructions.md
-
+src/components/Header/backups
