+++ src/components/Header/header.scss
@@ -11,9 +11,9 @@
   position: fixed;
   top: 20px;
   width: 100%;
-  @media(max-width:768px){
-
-  }
+  // @media(max-width:650px){
+    
+  // }
   &__left {
     display: flex;
     flex-grow: 2;
@@ -40,29 +40,54 @@
     min-width: 50%;
     width: auto;
     backdrop-filter: blur(25px);
+    
+
     @media(max-width:650px){
       height: 100vh;
-      width:17* vars.$gutter ;
+      width:15* vars.$gutter ;
       flex-direction: column;
       display: flex;
       justify-content: flex-start;
       align-items: flex-end;
       margin-top: 2* vars.$gutter;
+      padding-left: 3* vars.$gutter;
+      position: relative;
     }
-    &--close {
-      visibility: hidden;
-      @media(max-width:650px){
-        visibility: visible;
-        background-color: 0;
-        font-size: vars.$gutter;
-        background:none;
-        display: block;
-        color: vars.$secondaryColor;
-        background-color: none;
-        width:1.5* vars.$gutter;
-        height: 1.5* vars.$gutter;
-      }
-    }
+    // &--close {
+    //   visibility: hidden;
+    //   @media(max-width:650px){
+    //     visibility: visible;
+    //     background-color: 0;
+    //     font-size: vars.$gutter;
+    //     background:none;
+    //     display: block;
+    //     color: vars.$secondaryColor;
+    //     background-color: none;
+    //     width:1.5* vars.$gutter;
+    //     height: 1.5* vars.$gutter;
+    //     position: absolute;
+    //     top:0;
+    //     right: 10px;
+    //   }
+      
+    // }
+    // &--open {
+    //   visibility: hidden;
+    //   @media(max-width:650px){
+    //     visibility: visible;
+    //     background-color: 0;
+    //     font-size: vars.$gutter;
+    //     background:none;
+    //     display: block;
+    //     color: vars.$secondaryColor;
+    //     background-color: none;
+    //     width:1.5* vars.$gutter;
+    //     height: 1.5* vars.$gutter;
+    //     position: fixed;
+    //     top:60px;
+    //     right: 10px;
+    //   }
+    // }
     //ul
     &__menu {
       display: flex;
