+++ src/components/Header/navbarDesktop.scss
@@ -1,0 +1,100 @@
+@use '../../styles/vars';
+@use '../../styles/reset';
+
+.header {
+  // display: flex;
+  // justify-content: space-between;
+  // font-size: 16px;
+  // letter-spacing: 2.7px;
+  // font-family: 'Barlow Condensed', sans-serif;
+  // color: vars.$mainColor;
+  // position: fixed;
+  // top: 20px;
+  // width: 100%;
+  @media(max-width:650px){
+    
+  }
+
+  &__nav {
+    // background-color: red;
+    // display: flex;
+    // justify-content: center;
+    // height:4* vars.$gutter;
+    // min-width: 50%;
+    // width: auto;
+    // backdrop-filter: blur(25px);
+    @media(max-width:768px){
+      // background-color: aqua;
+      // backdrop-filter: blur(1px);
+    }
+    // @media(max-width:650px){
+    //   height: 100vh;
+    //   width:17* vars.$gutter ;
+    //   flex-direction: column;
+    //   display: flex;
+    //   justify-content: flex-start;
+    //   align-items: flex-end;
+    //   margin-top: 2* vars.$gutter;
+    //   padding-left: 3* vars.$gutter;
+    //   position: relative;
+    // }
+    //ul
+    &__menu {
+      display: flex;
+      justify-content: space-around;
+      align-items: center;
+      width: 100%; 
+      @media(max-width:650px){
+        flex-direction: column;
+        align-items: flex-start;
+        justify-content: flex-start;
+        margin-top:2* vars.$gutter;
+      }
+      &--li{
+        display: flex;
+      }
+      &--item {
+        margin-right: vars.$gutter;
+        font-size: 1rem;
+        text-transform: uppercase;
+        @media(max-width:650px){
+          // display: none;
+          // text-align: left;
+          // margin-right: 0;
+          // margin-top:2* vars.$gutter;
+          // width: 100%;
+        }
+        &:hover {
+        text-decoration: underline;
+        text-underline-offset: 1.5* vars.$gutter;
+        text-decoration-color: grey;
+        text-decoration-thickness: 3px;
+        @media(max-width:650px){
+          // text-decoration: none;
+        }
+        }
+      }
+    }
+  }
+}
+.span {
+  font-weight: bolder;
+  @media(max-width:768px){
+    display: none;
+  }
+  // @media(max-width:650px){
+  //   display:flex;
+  //   flex-wrap: nowrap;
+  // }
+}
+  .active {
+    text-decoration: underline;
+    text-underline-offset: 1.5* vars.$gutter;
+    text-decoration-thickness: 3px;
+    // @media(max-width:650px){
+    //   text-decoration: none;
+    // }
+  }
+  pre {
+    display: flex;
+  }
