+++ src/components/Header/navbarMobile.scss
@@ -1,0 +1,152 @@
+@use '../../styles/vars';
+@use '../../styles/reset';
+
+.header {
+  // display: none;
+//   display: flex;
+//   justify-content: space-between;
+//   font-size: 16px;
+//   letter-spacing: 2.7px;
+//   font-family: 'Barlow Condensed', sans-serif;
+//   color: vars.$mainColor;
+//   position: fixed;
+//   top: 20px;
+//   width: 100%;
+//   @media(max-width:650px){
+    
+//   }
+  // .header__left {
+  //   display: flex;
+  //   flex-grow: 2;
+  //   padding-right:2* vars.$gutter;
+
+  //   &__logo {
+  //     padding: vars.$gutter;
+  //   }
+  //   &__divider {
+  //     padding: vars.$gutter;
+  //     padding-right: 2* vars.$gutter;
+  //     width:100%;
+  //     height: 10px;
+  //     border-bottom: 1px solid vars.$secondaryColor;
+  //     @media(max-width:768px){
+  //       display: none;
+  //     }
+  //   }
+  // }
+  &__nav {
+    // display: none;
+    // background-color: aqua;
+    // display: flex;
+    // justify-content: center;
+    // height:4* vars.$gutter;
+    // min-width: 50%;
+    // width: auto;
+    // backdrop-filter: blur(25px);
+
+    @media(max-width:650px){
+      height: 100vh;
+      width:17* vars.$gutter ;
+      flex-direction: column;
+      display: flex;
+      justify-content: flex-start;
+      align-items: flex-end;
+      margin-top: 2* vars.$gutter;
+      padding-left: 3* vars.$gutter;
+      position: relative;
+    }
+    &--close {
+      visibility: hidden;
+      @media(max-width:650px){
+        visibility: visible;
+        background-color: 0;
+        font-size: vars.$gutter;
+        background:none;
+        display: block;
+        color: vars.$secondaryColor;
+        background-color: none;
+        width:1.5* vars.$gutter;
+        height: 1.5* vars.$gutter;
+        position: absolute;
+        top:0;
+        right: 10px;
+      }
+      
+    }
+    &--open {
+      visibility: hidden;
+      @media(max-width:650px){
+        visibility: visible;
+        background-color: 0;
+        font-size: vars.$gutter;
+        background:none;
+        display: block;
+        color: vars.$secondaryColor;
+        background-color: none;
+        width:1.5* vars.$gutter;
+        height: 1.5* vars.$gutter;
+        position: fixed;
+        top:60px;
+        right: 10px;
+      }
+    }
+    //ul
+    &__menu {
+      display: flex;
+      justify-content: space-around;
+      align-items: center;
+      width: 100%; 
+      @media(max-width:650px){
+        flex-direction: column;
+        align-items: flex-start;
+        justify-content: flex-start;
+        margin-top:2* vars.$gutter;
+      }
+      &--li{
+        display: flex;
+      }
+      &--item {
+        margin-right: vars.$gutter;
+        font-size: 1rem;
+        text-transform: uppercase;
+        @media(max-width:650px){
+          text-align: left;
+          margin-right: 0;
+          margin-top:2* vars.$gutter;
+          width: 100%;
+        }
+        &:hover {
+        text-decoration: underline;
+        text-underline-offset: 1.5* vars.$gutter;
+        text-decoration-color: grey;
+        text-decoration-thickness: 3px;
+        @media(max-width:650px){
+          text-decoration: none;
+        }
+        }
+      }
+    }
+  }
+
+.span {
+  font-weight: bolder;
+  @media(max-width:768px){
+    display: none;
+  }
+  @media(max-width:650px){
+    display:flex;
+    flex-wrap: nowrap;
+  }
+}
+  .active {
+    text-decoration: underline;
+    text-underline-offset: 1.5* vars.$gutter;
+    text-decoration-thickness: 3px;
+    @media(max-width:650px){
+      text-decoration: none;
+    }
+  }
+  pre {
+    display: flex;
+  }
+}
